split flamechart-view-state into reducer and container
